% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpoints.R
\name{ggpoints}
\alias{ggpoints}
\title{ggpoints}
\usage{
ggpoints(
  data = NULL,
  x,
  y,
  annotations = c("lm", "cor"),
  linetypes = c("xy.line", "lm.line"),
  cor.method = "spearman",
  show.cor.pvalue = TRUE,
  label.x.cor = 0.01,
  label.y.cor = 0.9,
  label.x.npc = "right",
  label.y.npc = "top",
  point.pch = 21,
  point.size = 2,
  point.fill = "#3b518be5",
  point.col = "#3b518be5",
  xy.line.col = "grey50",
  lm.line.col = "blue",
  lm.line.size = 1.1
)
}
\arguments{
\item{data:}{data frame: if non, using x and y to plot.}

\item{x:}{x}

\item{y:}{y}

\item{annotations:}{annotations text in the plot.}

\item{linetypes:}{add lines into the plot.}

\item{label.x.cor:}{label.x.cor of cor annotation text}

\item{label.y.cor:}{label.y.cor of cor annotation text}

\item{label.x.npc:}{label.x.npc of lm annotation text}

\item{label.y.npc:}{label.x.npc of lm annotation text}
}
\description{
plot the x and y points using ggplot. The annotations can be correlation coefficient and linear regression function.
}
\details{
The types of annotation:

3. cor: add cor.test of R and p-value.

4. lm: lm information, y ~ x and R^2.

The types of linetypes:

1. xy.line: add y = x dashed lines.

2. lm.line: add y ~ x lines.
}
\examples{
#data
set.seed(100)
x = rnorm(20)
y = x*1.1 + 1 + rnorm(20, mean = 0.01)
data = data.frame(x = x, y = y)

#plot all data
ggpoints(x = x, y = y)
ggpoints(data = data, x = "x", y = "y")

#plot lines
ggpoints(x = x, y = y, annotations = c("lm"), linetypes = "lm.line" )

#adjust the position of label
ggpoints(x = x, y = y, annotations = c("lm"),linetypes = "lm.line", label.x.npc = 0.1, label.y.npc = 0.9)


}
