% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftOverData.R
\name{liftOverData}
\alias{liftOverData}
\title{liftOverData}
\usage{
liftOverData(data, chainFiles = "hg38ToHg19.over.chain", chrVersion = "hg19")
}
\arguments{
\item{data:}{data frame: Three columns are required, including Chromosome, Start and End}

\item{chainFiles:}{liftover files. see details.}

\item{chrVersion:}{chr version of the output.}
}
\description{
The liftOver facilities developed in conjunction with the UCSC browser track infrastructure are available for transforming data in GRanges format
}
\details{
chainFiles can be downloaded from the following urls:

- hg19tohg38: http://hgdownload.soe.ucsc.edu/goldenPath/hg19/liftOver/hg19ToHg38.over.chain.gz

- hg38tohg19: http://hgdownload.soe.ucsc.edu/goldenPath/hg38/liftOver/hg38ToHg19.over.chain.gz
}
\examples{
#load examples data.

data("ASCAT_hg38_examples")

hg38ToHg19.over.chain = system.file(package="Jerry", "extdata", "hg38ToHg19.over.chain")

# Three columns, including Chromosome, Start and End, are required to liftover.
head(stad_ascat)

stad_ascat_hg19 = liftOverData(data = stad_ascat,
                              chainFiles = hg38ToHg19.over.chain,
                              chrVersion = "hg19")
head(stad_ascat_hg19)

}
