% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{plot_survival}
\usage{
plot_survival(
  data,
  variables,
  cutoff,
  time = "time",
  status = "status",
  palette = NULL,
  minprop = 0.4,
  return.data = FALSE,
  ...
)
}
\arguments{
\item{data}{the input data.}

\item{variables}{the variable to be explored.}

\item{cutoff}{cutoff methods. cutoff methods is setting either "median", "mean", "best" or the certain values. When the variables is numerical, we divide the variables into two groups according to the inputed values.}

\item{time}{columns names specifying time.}

\item{status}{columns names specifying status.}

\item{palette}{colors for different groups.}

\item{minprop}{Setting minprop values if cutoff methods is "best"}

\item{return.data}{whether return the data frame, Default is FALSE.}

\item{...}{parameters from survminer::ggsurvplot() to control the styles of figures}
}
\description{
The complex survival plot.
}
\examples{
library(survival)
library(ggsurvfit)
library(gtsummary)
library(ezcox)

head(lung )

#get HR values.
zz = lung |>
  mutate(
    status = recode(status, `1` = 0, `2` = 1),
    time = time
  ) |>
  ezcox(
    covariates = c("pat.karno", "meal.cal", "wt.loss"),
    controls = c("age", "sex"),
    time = "time",
    status = "status",
    return_models = TRUE
  )

mds = get_models(zz)

show_models(mds, drop_controls = TRUE,  merge_models = T)

#Table of summary
tbl_regression(mds$`Surv ~ pat.karno + age + sex` , exp = TRUE)

#Plot Survival
surv.list = plot_survival(
lung |>
  mutate(
    status = recode(status, `1` = 0, `2` = 1),
    time = time
  ),
variables = c("pat.karno", "age"),
time = "time",
status = "status",
legend.labs = c("High","Low"),
cutoff = "best",
minprop = 0.2,
return.data = T
)

surv.list$plot.list

#The median survival days.
gtsummary::tbl_survfit(
  surv.list$surv.fit$pat.karno.g,
 probs  = 0.5,
  label_header = "**Median survival (95\% CI)**"
)

}
