% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_nominal.R
\name{plot_survival_nominal}
\alias{plot_survival_nominal}
\title{plot_survival_nominal}
\usage{
plot_survival_nominal(
  data,
  variables,
  time = "time",
  status = "status",
  palette = NULL,
  return.data = TRUE,
  ...
)
}
\arguments{
\item{data}{the input data.}

\item{variables}{the variable to be explored.}

\item{time}{columns names specifying time.}

\item{status}{columns names specifying status.}

\item{palette}{colors for different groups.}

\item{return.data}{whether return the data frame, Default is FALSE.}

\item{...}{parameters from survminer::ggsurvplot() to control the styles of figures}

\item{cutoff}{cutoff methods. cutoff methods is setting either "median", "mean", "best" or the certain values. When the variables is numerical, we divide the variables into two groups according to the inputed values.}

\item{minprop}{Setting minprop values if cutoff methods is "best"}
}
\description{
Plot survival for continuous variables. The continuous variables were divided into high and low groups by setting methods including "median", "mean" or "best".
}
\examples{
library(survival)
library(ggsurvfit)
library(gtsummary)
library(ezcox)

head(lung )
plot_survival_nominal(lung, variables = "ph.ecog")

}
