% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_colors.R
\name{sci_colors}
\alias{sci_colors}
\title{Color setting}
\usage{
sci_colors(n = 1, plot = F)
}
\arguments{
\item{n}{number of colors to select. A total of 36 colors are saved.}

\item{rev}{reverse output the colors.}

\item{random}{random select number of colors.}
}
\description{
Color setting
}
