% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival_nominal.R
\name{plot_survival_nominal}
\alias{plot_survival_nominal}
\title{plot_survival_nominal}
\usage{
plot_survival_nominal(
  data,
  variables,
  time = "time",
  status = "status",
  palette = NULL,
  ylab = "Overall Survival (OS)",
  return.data = TRUE,
  ...
)
}
\arguments{
\item{data}{the input data.}

\item{variables}{the variable to be explored.}

\item{time}{columns names specifying time.}

\item{status}{columns names specifying status.}

\item{palette}{colors for different groups.}

\item{ylab}{Default: Overall Survival (OS)}

\item{return.data}{whether return the data frame, Default is FALSE.}

\item{...}{parameters from survminer::ggsurvplot() to control the styles of figures}
}
\description{
Plot survival for nominal variables.
}
\examples{
library(survival)
library(ggsurvfit)
library(gtsummary)
library(ezcox)

head(lung )
plot_survival_nominal(lung, variables = "ph.ecog")

}
